requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# PDF processing
pypdf2>=3.0.1

langchain>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
streamlit>=1.31.0
pandas>=2.2.0
numpy>=2.0.0
sentence-transformers>=3.0.0
tiktoken>=0.7.0

# Later we'll add more